requirements.txt
Django==4.2.7
python-decouple==3.8
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
django-extensions==3.2.3
